requirements.txt
dj-database-url==0.5.0
Django==3.0.8
django-sekizai==1.1.0
Pillow==7.2.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
tzlocal==2.1
Unidecode==1.1.1
django-ckeditor==6.0.0